+++ includes/Core/class-plugin.php
@@ -33,7 +33,7 @@ use function do_action;
  */
 final class Plugin {
 
-        public const VERSION = '1.0.0';
+        public const VERSION = '1.0.6';
 
 
 	/**